% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_herring.R
\name{clean_herring_data}
\alias{clean_herring_data}
\title{Cleans herring data}
\usage{
clean_herring_data(df)
}
\arguments{
\item{df}{Takes in the a dataset}
}
\value{
A cleaned dataset
}
\description{
cleans already read dataset
}
