% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_herring.R
\name{create_leaflet_map}
\alias{create_leaflet_map}
\title{Create Leaflet map of Barents Sea stations}
\usage{
create_leaflet_map(station_data)
}
\arguments{
\item{station_data}{Aggregated station data}
}
\value{
A Leaflet map object
}
\description{
Generates a Leaflet map with station markers and hover labels.
}
