% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_herring.R
\name{herring_server}
\alias{herring_server}
\title{Shiny server logic for herring app}
\usage{
herring_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Handles reactivity, map clicks, and plotting.
}
