% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_herring.R
\name{plot_length_hist}
\alias{plot_length_hist}
\title{plots length histogram}
\usage{
plot_length_hist(df, station_id)
}
\arguments{
\item{df}{dataframe with excel data}

\item{station_id}{For station number in original file}
}
\value{
plot for length histogram
}
\description{
plots the aesthetics and the displayed bins for the plot
}
