% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_herring.R
\name{plot_station_histograms}
\alias{plot_station_histograms}
\title{Create histograms for selected station}
\usage{
plot_station_histograms(data, station_id)
}
\arguments{
\item{data}{Full herring dataset}

\item{station_id}{Selected station number}
}
\value{
A list with ggplot objects
}
\description{
Generates histograms of length and weight for a given station.
}
