% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_herring.R
\name{read_herring_data}
\alias{read_herring_data}
\title{Read Herring}
\usage{
read_herring_data(path)
}
\arguments{
\item{path}{Takes in the path of dataset}
}
\value{
A shiny app object

A read dataset
}
\description{
Reads in excel dataset
}
\details{
Launches the "hansensherring" shiny app

This function launches the Hansen's Herring shiny app.
This app displays a Leaflet-map over the western part of Svalbard Archipelago,
with 20 generated trawl stations where herring *C.harengus* is found.
catch total, mean weigth and mean length is displayed while hovering over
stations, while clicking them displays histograms showing weight and length.
}
