% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_herring.R
\name{run_hansensherring}
\alias{run_hansensherring}
\title{launches the server and the app visuals}
\usage{
run_hansensherring()
}
\arguments{
\item{input}{data from earlier functions}

\item{output}{final output for shiny app}

\item{session}{for displaying shiny app in seperate window}
}
\value{
shiny application
}
\description{
launches the shiny server and the accompanying visuals
}
