% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_herring.R
\name{summarize_stations}
\alias{summarize_stations}
\title{Summarizes stations}
\usage{
summarize_stations(df)
}
\arguments{
\item{df}{Takes in the path}
}
\value{
mean measurements and fish amounts grouped by station and location
}
\description{
groups and calculate mean of fish measurements
}
